research/requirements.txt
alpaca-py>=0.30,<0.40
pandas>=2.1,<3
numpy>=1.26,<2
pyarrow>=14.0
python-dotenv>=1.0
scikit-learn>=1.4,<2
hmmlearn>=0.3,<0.4
hdbscan>=0.8.33,<0.9
lightgbm>=4.3,<5
joblib>=1.3
